Detect deprecated ISO language codes
Also moved LangMap handling code into own module.
